Refactor Android review reaction overlay
