+++ apps/android/feature/review/src/main/java/com/flashcardsopensourceapp/feature/review/ReviewReactionPalette.kt
@@ -1,0 +1,11 @@
+package com.flashcardsopensourceapp.feature.review
+
+import androidx.compose.ui.graphics.Color
+
+internal val reviewReactionRedColor: Color = Color(red = 0.96f, green = 0.13f, blue = 0.17f)
+internal val reviewReactionOrangeColor: Color = Color(red = 1.00f, green = 0.50f, blue = 0.10f)
+internal val reviewReactionYellowColor: Color = Color(red = 1.00f, green = 0.82f, blue = 0.14f)
+internal val reviewReactionGreenColor: Color = Color(red = 0.12f, green = 0.76f, blue = 0.34f)
+internal val reviewReactionBlueColor: Color = Color(red = 0.16f, green = 0.62f, blue = 1.00f)
+internal val reviewReactionPurpleColor: Color = Color(red = 0.58f, green = 0.34f, blue = 0.98f)
+internal val reviewReactionPinkColor: Color = Color(red = 1.00f, green = 0.32f, blue = 0.68f)
